requirements.txt
python-binance
openpyxl